Fix version code type to int

+++ {{cookiecutter.repo_name}}/config/src/main/java/{{cookiecutter.package_name_dir}}/config/data/Config.kt
@@ -1,5 +1,5 @@
 package {{ cookiecutter.package_name }}.config.data
 
 data class Config(
-    val minSupportedVersionCode: Long,
+    val minSupportedVersionCode: Int,
 )
